+++ src/AnomalyDetectors/Loda.php
@@ -11,6 +11,7 @@ use Rubix\Tensor\Matrix;
 use Rubix\Tensor\Vector;
 use Rubix\ML\Persistable;
 use Rubix\ML\Datasets\Dataset;
+use Rubix\ML\Other\Traits\RankSingle;
 use Rubix\ML\Other\Traits\PredictsSingle;
 use Rubix\ML\Other\Specifications\DatasetIsCompatibleWithEstimator;
 use InvalidArgumentException;
@@ -38,7 +39,7 @@ use const Rubix\ML\LOG_EPSILON;
  */
 class Loda implements Estimator, Learner, Online, Ranking, Persistable
 {
-    use PredictsSingle;
+    use PredictsSingle, RankSingle;
     
     protected const MIN_SPARSE_DIMENSIONS = 3;